feat: context added to all i/o functions

+++ internal/services/authService.go
@@ -16,7 +16,7 @@ func NewAuthService() AuthService {
 }
 
 type Claims struct {
-	Sub int       `json:"sub"`
+	Sub int64     `json:"sub"`
 	Exp time.Time `json:"exp"`
 	jwt.RegisteredClaims
 }
